Version 70.   ***WIRING CHANGE***
This changes the pins for the ZProbe wiring to allow the ZProbe
to be used with the Duet and with the Duet+expansion board without
changing the firmware.  The new Duet expansion header pins for
the Z Probe are:

pin 2 >  .........................   < pin 50
pin 1 >  ...................ST....   < pin 49

S is the ZProbe signal output.  T is the trigger/modulator input.  S goes
to Arduino pin A10 and T goes to D52.