Intermediate version 0.77a
M105 now only reports the temps of active heaters.
The first tool defined is automatically selected as the active one.

Automatic tool selection is fine for single-head machines as there will in
general only be one tool (though you are free to define more that, for example,
operate at different temperatures).  But this automatic selection will go away
in future releases and tools will have to be explicitly selected after definition.

It is not a good idea to select tools in config.g because tool selection may cause
the RepRap to move or to switch heaters on.  config.g shouldn't do either of these.
It is better to do tool selection in the header G Codes of printable G Code files.