Version 0.78
This is making the dev branch identical to the duet branch.  From now on the
dev branch will be for what its name implies: the experimental branch for the
master.

The following branches will probably be removed from the repository in
the near future:

  due
  duet
  duet-spi
  vel

And a new branch - test - will be added immediately.  test will be derived
from dev when a version is to be tested for using to upgrade the master.  When
test has passed the release tests for the firmware then that will become
the master.