Version 0.86
IMPORTANT: with this version the direction of the Ormerod X axis is
deliberately reversed.  This means that the colour ordering of the
X axis wires on the Duet board is now the same as that for all the
other axes.

If you upgrade to this and subsequent versions from Version 0.85 or
a previous one, you will need to swap your X wires.

This upgrade also implements the M569 (Set axis direction values)
M code, so you can set the logical values that decide all axis directions.