Version 0.89
Fixed bug whereby low feeds on their own (especially for the extruder) were
causing the wrong feedrate.

This was caused when the acceleration was fast enough to get to the speed requested
in a single step.  The Euler integration used to compute velocities is only
approximate.  This approximation is good over a reasonable number of steps.  But at
just one step it can give an answer that is a long way out.