Version 0.78a
Release for Ormerods 1 and 2.
Separate SD Image directories creted for each Ormerod model.
Axis direction by M569 P0 Sb implemented, and used in the config.g for the Ormerod 2 X axis.
Setting the bed equation reports the probe points to the USB and the web Messages
The machine's name is used in the web interface.