Version 1.05 ------------
Bug in reporting tempereatures to Pronterface fixed.  Also
temperatures will continue to be reported even when a heater
develops a fault.

Made a start on automating the setup of the machine.  The SD-image
folder now contains setups for all the main models and variants and
a README on which files to copy and how.

These variants need to be tested.