Version 1.06 ------------
A whole bunch of G and M codes have been categorized as 'wait for the
move buffer to exhaust before doing this'.  Things like changing the IP
address.  Now, it would be eccentric to change the IP address in the middle
of a print file.  But, if you do so, you probably want the move buffer
to be exhausted before you do.