Version 1.04 ------------
This is the release of version 1.04 into the master branch.

This release has a web interface that supports multiple heaters and drives
to support multiple extruders.

The movement code has been improved to make moves smoother.

Now all G and M codes for which it makes sense report their parameters
back when called with no arguments.

There ehave been a number of internal improvements, simplifications
and enhancements.